recipe/src/main/java/dw/recipe/controllers/IndexController.java
package dw.recipe.controllers;

import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.RequestMapping;

import dw.recipe.services.RecipeService;
import lombok.extern.slf4j.Slf4j;

@Controller
@Slf4j
public class IndexController {

	private final RecipeService recipeService;
	
	public IndexController(RecipeService recipeService) {
		super();
		this.recipeService = recipeService;
	}

	@RequestMapping({"","/","index","/index"})
	public String getIndex(Model model) {
		model.addAttribute("recipes", recipeService.getRecipes());
		
		return "index";
	}
	
}
